+++ .github/workflows/selfdrive_tests.yaml
@@ -299,6 +299,7 @@ jobs:
                         $UNIT_TEST selfdrive/loggerd && \
                         $UNIT_TEST selfdrive/car && \
                         $UNIT_TEST selfdrive/locationd && \
+                        selfdrive/locationd/test/_test_locationd_lib.py && \
                         $UNIT_TEST selfdrive/athena && \
                         $UNIT_TEST selfdrive/thermald && \
                         $UNIT_TEST selfdrive/hardware/tici && \
@@ -364,7 +365,7 @@ jobs:
                         CI=1 AZURE_TOKEN='$AZURE_TOKEN' python selfdrive/test/process_replay/test_processes.py -j$(nproc) --upload-only"
     - name: "Upload coverage to Codecov"
       uses: codecov/codecov-action@v2
-  
+
   model_replay_onnx:
     name: model replay onnx
     runs-on: ubuntu-20.04
